+++ litmus-portal/frontend/package.json
@@ -92,7 +92,7 @@
     "@typescript-eslint/eslint-plugin": "^3.5.0",
     "@typescript-eslint/parser": "^3.5.0",
     "cross-env": "^5.2.0",
-    "cypress": "^4.12.1",
+    "cypress": "^5.1.0",
     "cypress-react-unit-test": "^4.11.2",
     "eslint": "^6.8.0",
     "eslint-config-airbnb": "^18.2.0",
